Add files via upload
通过一些内置整数处理方案，已经能够正常agc，但是代码还需整理优化

+++ WebRtcAgc_InitDigital.m
@@ -1,12 +1,12 @@
 function [y, stt] = WebRtcAgc_InitDigital(stt, agcMode, param)
     if (agcMode == param.kAgcModeFixedDigital)
     %        start at minimum to find correct gain faster
-        stt.capacitorSlow = 0;
+        stt.capacitorSlow = int32(0);
     else
     %       start out with 0 dB gain
-        stt.capacitorSlow = 134217728; 
+        stt.capacitorSlow = int32(134217728); 
     end
-        stt.capacitorFast = 0;
+        stt.capacitorFast = int32(0);
         stt.gain = 65536;
         stt.gatePrevious = 0;
         stt.agcMode = agcMode;
